requirements.txt

pandas
dotenv
numpy
matplotlib
scikit-learn
huggingface_hub
cerebras-cloud-sdk
datasets
jinja2